Move check_(lists,function_tensors)_enabled into experimental.py
