+++ guppylang/cfg/builder.py
@@ -16,6 +16,7 @@ from guppylang.cfg.bb import BB, BBStatement
 from guppylang.cfg.cfg import CFG
 from guppylang.checker.core import Globals
 from guppylang.error import GuppyError, InternalGuppyError
+from guppylang.experimental import check_lists_enabled
 from guppylang.nodes import (
     DesugaredGenerator,
     DesugaredListComp,
@@ -26,7 +27,6 @@ from guppylang.nodes import (
     NestedFunctionDef,
     PyExpr,
 )
-from guppylang.tys.builtin import check_lists_enabled
 from guppylang.tys.ty import NoneType
 
 # In order to build expressions, need an endless stream of unique temporary variables
